efi: Add an EFI signature blob parser
mainline inclusion
from mainline-4.20-rc7
commit 0bc9ae395b3f3b6557f0c5f0a0b0cd2fd5c00a04
Category: Bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I5VB21
CVE: CVE-2020-26541
Reference: https://cwe.mitre.org/data/definitions/347.html

Add a function to parse an EFI signature blob looking for elements of
interest. A list is made up of a series of sublists, where all the
elements in a sublist are of the same type, but sublists can be of
different types.

For each sublist encountered, the function pointed to by the
get_handler_for_guid argument is called with the type specifier GUID and
returns either a pointer to a function to handle elements of that type or
NULL if the type is not of interest.

If the sublist is of interest, each element is passed to the handler
function in turn.

K2CI-Arch: All
Signed-off-by: lijun <lijun01@kylinos.cn>
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Nayna Jain <nayna@linux.ibm.com>
Acked-by: Serge Hallyn <serge@hallyn.com>
Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>
Change-Id: Iba8db82bf1ef14cd1ffa55c64bba6bf28ae9e298