vfio: return "keepalive" flag in ioctl VFIO_DEVICE_GET_INFO
When user land vfio driver is restarted, its underlying device may
has been in "keepalive" state. The restarted vfio driver must not
do any operation to clobber hardware states. The vfio driver need
to be notified of this state.

Add a new VFIO device flag VFIO_DEVICE_FLAGS_KEEPALIVE to indicate
that this device is in "keepalive" state. This flag is returned to
user land when ioctl VFIO_DEVICE_GET_INFO is issued against this
device.

Signed-off-by: Jason Zeng <jason.zeng@intel.com>