PCI: Support ASpeed VGA cards behind a misbehaving bridge
According to PCI-to-PCI bridge spec, bit 3 of Bridge Control Register
is VGA Enable bit which modifies the response by the bridge to VGA
compatible addresses.

The Bridge Control register provides extensions to the Command register
that are specific to a bridge. The Bridge Control register provides
many of the same controls for the secondary interface that are provided
by the Command register for the primary interface. There are some bits
that affect the operation of both interfaces of the bridge.

If the VGA Enable bit is set, the bridge will decode and forward the
following accesses on the primary interface to the secondary interface.

Forwarding of these accesses is qualified by the I/O Enable and
Memory Enable bits in the Command register, and VGA Enable bit modifies
the response by the bridge to VGA compatible addresses.

when 0: do not forward VGA compatible memory and I/O addresses from
        the primary to secondary interface (addresses defined below)
        unless they are enabled for forwarding by the defined I/O
        and memory address ranges

when 1: forward VGA compatible memory and I/O addresses (addresses
        defined below) from the primary interface to the secondary
        interface (if the I/O Enable and Memory Enable bits are set)
        independent of the I/O and memory address ranges and
        independent of the ISA Enable bit

* memory accesses in the range 000A 0000h to 000B FFFFh

* I/O addresses in the first 64 KB of the I/O address space
  (AD[31:16] are 0000h) where AD[9:: 0] are in the ranges
  3B0h to 3BBh and 3C0h to 3DFh (inclusive of ISA address
  aliases - AD[15::10] are not decoded)

If the VGA Enable bit is set, forwarding of these accesses is
independent of the I/O address range and memory address ranges
defined by the I/O Base and Limit registers, the Memory Base
and Limit registers, and the Prefetchable Memory Base and Limit
registers of the bridge.

Forwarding of these accesses is also independent of the settings
of the ISA Enable bit (in the Bridge Control register) or VGA
Palette Snoop bits (in the Command register),

The AST2500 hardward does not set the VGA Enable bit on its bridge
control register, which causes vgaarb subsystem don't think the VGA
card behind this bridge as a valid boot vga device.

So we provide a quirk to fix Xorg auto-detection.

See similar bug:

https://patchwork.kernel.org/project/linux-pci/patch/20170619023528.11532-1-dja@axtens.net/

Change-Id: I18b5a7f232d718658cbeafbb7ac36539de3e76f3
Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>
Signed-off-by: Jingfeng Sui <suijingfeng@loongson.cn>