Intel: iommu/vt-d: Delegate the identity domain to upper layer commit 4de354ec
 upstream
Backport summary: Backport to kernel 4.19.57 to fix kernel panic for
intel_iommu=on while passthrough disable on ICX.

This allows the iommu generic layer to allocate an identity domain
and attach it to a device. Hence, the identity domain is delegated
to upper layer. As a side effect, iommu_identity_mapping can't be
used to check the existence of identity domains any more.
Signed-off-by: default avatarJames Sewart <jamessewart@arista.com>
Signed-off-by: default avatarLu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: default avatarJoerg Roedel <jroedel@suse.de>
Signed-off-by: default avatarYouquan Song <youquan.song@intel.com>
Signed-off-by: Xibo.Wang <wangxb12@chinatelecom.cn>