Intel: iommu: Add API to request DMA domain for device commit 7423e017
 upstream
Backport summary: Backport to kernel 4.19.57 to fix kernel panic for
intel_iommu=on while passthrough disable on ICX.

Normally during iommu probing a device, a default doamin will
be allocated and attached to the device. The domain type of
the default domain is statically defined, which results in a
situation where the allocated default domain isn't suitable
for the device due to some limitations. We already have API
iommu_request_dm_for_dev() to replace a DMA domain with an
identity one. This adds iommu_request_dma_domain_for_dev()
to request a dma domain if an allocated identity domain isn't
suitable for the device in question.
Signed-off-by: default avatarLu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: default avatarJoerg Roedel <jroedel@suse.de>
Signed-off-by: default avatarYouquan Song <youquan.song@intel.com>
Signed-off-by: Xibo.Wang <wangxb12@chinatelecom.cn>