Intel: iommu/vt-d: Identify default domains replaced with private commit 942067f1
 upstream
Backport summary: Backport to kernel 4.19.57 to fix kernel panic for
intel_iommu=on while passthrough disable on ICX.

When we put a device into an iommu group, the group's default
domain will be attached to the device. There are some corner
cases where the type (identity or dma) of the default domain
doesn't work for the device and the request of a new default
domain results in failure (e.x. multiple devices have already
existed in the group). In order to be compatible with the past,
we used a private domain. Mark the private domains and disallow
some iommu apis (map/unmap/iova_to_phys) on them.
Signed-off-by: default avatarLu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: default avatarJoerg Roedel <jroedel@suse.de>
Signed-off-by: default avatarYouquan Song <youquan.song@intel.com>
Signed-off-by: Xibo.Wang <wangxb12@chinatelecom.cn>