Intel: iommu/vt-d: Handle 32bit device with identity default domain commit 98b2fffb
 upstream
Backport summary: Backport to kernel 4.19.57 to fix kernel panic for
intel_iommu=on while passthrough disable on ICX.

The iommu driver doesn't know whether the bit width of a PCI
device is sufficient for access to the whole system memory.
Hence, the driver checks this when the driver calls into the
dma APIs. If a device is using an identity domain, but the
bit width is less than the system requirement, we need to use
a dma domain instead. This also applies after we delegated
the domain life cycle management to the upper layer.
Signed-off-by: default avatarLu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: default avatarJoerg Roedel <jroedel@suse.de>
Signed-off-by: default avatarYouquan Song <youquan.song@intel.com>
Signed-off-by: Xibo.Wang <wangxb12@chinatelecom.cn>