Intel: iommu/vt-d: Cleanup get_valid_domain_for_dev() commit 4ec066c7
 upstream
Backport summary: Backport to kernel 4.19.57 to fix kernel panic for
intel_iommu=on while passthrough disable on ICX.

Previously, get_valid_domain_for_dev() is used to retrieve the
DMA domain which has been attached to the device or allocate one
if no domain has been attached yet. As we have delegated the DMA
domain management to upper layer, this function is used purely to
allocate a private DMA domain if the default domain doesn't work
for ths device. Cleanup the code for readability.
Signed-off-by: default avatarLu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: default avatarJoerg Roedel <jroedel@suse.de>
Signed-off-by: default avatarYouquan Song <youquan.song@intel.com>
Signed-off-by: Xibo.Wang <wangxb12@chinatelecom.cn>