iommu/vt-d: Fix dmar pte read access not set error commit 942067f1 upstream Backport summary: Backport to kernel 4.19.57 to fix kernel panic for intel_iommu=on while passthrough disable on ICX.
If the default DMA domain of a group doesn't fit a device, it
will still sit in the group but use a private identity domain.
When map/unmap/iova_to_phys come through iommu API, the driver
should still serve them, otherwise, other devices in the same
group will be impacted. Since identity domain has been mapped
with the whole available memory space and RMRRs, we don't need
to worry about the impact on it.

Link: https://www.spinics.net/lists/iommu/msg40416.html

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Reported-by: default avatarJerry Snitselaar <jsnitsel@redhat.com>
Fixes: 942067f1

 ("iommu/vt-d: Identify default domains replaced with private")
Cc: stable@vger.kernel.org # v5.3+
Signed-off-by: default avatarLu Baolu <baolu.lu@linux.intel.com>
Reviewed-by: default avatarJerry Snitselaar <jsnitsel@redhat.com>
Signed-off-by: Xibo.Wang <wangxb12@chinatelecom.cn>