rtc: ls2x: Add support for the Loongson-2K/LS7A RTC
This RTC module is integrated into the Loongson-2K SoC and the LS7A
bridge chip. This version is almost entirely rewritten to make use of
current kernel API, and it supports both ACPI and DT.

This patch also make CONFIG_RTC_DRV_EFI=m. The purpose of this is to
make rtc-ls2x device to /dev/rtc0.

Signed-off-by: Huacai Chen <chenhuacai@kernel.org>
Signed-off-by: WANG Xuerui <git@xen0n.name>
Signed-off-by: Binbin Zhou <zhoubinbin@loongson.cn>
Cc: Alexandre Belloni <alexandre.belloni@bootlin.com>
Cc: "Rafael J. Wysocki" <rafael@kernel.org>
Cc: devicetree@vger.kernel.org
Cc: linux-acpi@vger.kernel.org
Cc: linux-mips@vger.kernel.org
Cc: loongarch@lists.linux.dev
Signed-off-by: Ming Wang <wangming01@loongson.cn>
Change-Id: If950a9b210fd1ffb32a2d4d7f0ad6c1fb5058303