[openEuler-1.0-LTS] Add sysfs file for overlayfs
In the actual production environment, Overlayfs often occurs some
problems, such as continuous growth of quota statistics. There is
no good way to define this category for the current environment.
Overlayfs has many restrictions in the actual use process but
it has not directly restricted at the code level. Therefore, some
strange problems introduced in operation errors will occur during
the use process. It is time-consuming. Based on this background,
Overlayfs's maintenance function needs to be enhanced.

This patch add sysfs file for overlayfs to recording the mount
information. The sysfs file will be deleted only when Overlayfs is truly
unloaded.

Signed-off-by: hiyoyolumi <862127279@qq.com>