Revert "ACPI / PM: LPIT: Register sysfs attributes based on FADT"
mainline inclusion
from mainline-v6.0-rc1
commit 403dbe3a5383d1e1bbcbe03d9fa6df0216e26102
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I612GQ
CVE: NA

--------------------------------

Revert commit 1cdda9486f51 ("ACPI / PM: LPIT: Register sysfs attributes
based on FADT"), because what it did was more confusing than it would
be to allow the sysfs attributes in question to be created regardless
of whether or not the relevant flag was set in the FADT.

If ACPI_FADT_LOW_POWER_S0 is not set, it need not mean that LPIT is
invalid and low-power S0 idle is not usable.  It merely means that
using S3 on the given system is more beneficial from the energy
saving perspective than using low-power S0 idle.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-by: Mario Limonciello <mario.limonciello@amd.com>
Signed-off-by: Zhao Lei <zhao_lei1@hoperun.com>