openEuler inclusion category: feature bugzilla: https://gitee.com/openeuler/kernel/issues/I66LJ2 reference: https://lore.kernel.org/lkml/20120816013623.872.94338.stgit@warthog.procyon.org.uk/ CVE: NA
--------------------------------

Add support for extracting PGP public key in Public-Subkey packet.
But it should be noted that the subkey needs to be wrapped into
a certificate contains a public subkey packet and a followed user
ID packet.

The reason why it is implemented in this way is that now the different
keys are parsed in different contexts. And the key description is
generated by the user ID packet. For PGP certificates, only the main
key packet has a followed user ID packet. So the subkey will fail to
be parsed because of the missing description.

The original patch (reference above) was also support like this. But
this is a transitional implementation. Because IMA Digest List feature
need to support to verify signature with public subkey. A better
implement will be committed in the feature.

Signed-off-by: Huaxin Lu <luhuaxin1@huawei.com>