usb: gadget: configfs: clean bind_deactivated flag.
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I671W8

-----------------------------------------------------------

After the USB device functions as the UVC device, the device cannot be
identified after being switched to the virtual network port, serial port,
or USB flash drive.

The bind_deactivated flag is set when the USB is used as the UVC.
However, the bind_deactivated flag is not set during uninstallation.
The USB device cannot be identified when it is used as another device.

Signed-off-by: taolan <taolan@huawei.com>