LoongArch: Export symbol for function smp_send_reschedule()
From upstream commit 143d64bdbdb85787953a70332f9e5f658b678550

Function smp_send_reschedule() is standard kernel API, which is defined
in header file include/linux/smp.h. However, on LoongArch it is defined
as an inline function, this is confusing and kernel modules can not use
this function.

Now we define smp_send_reschedule() as a general function, and add a
EXPORT_SYMBOL_GPL on this function, so that kernel modules can use it.

Signed-off-by: Bibo Mao <maobibo@loongson.cn>
Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>
Change-Id: Iafc643ec9282b60ad211f86d66804f30f9932304