arch: fixed pointer access before check in destroy_queue()
zhaoxin inclusion
category: bugfix
bugzilla: NA
CVE: NA

Reference: N/A

----------------------------------------------------------------

Since the qi pointer is accessed first in the function and then checked to see if it is NULL, an error could occur if qi is NULL anyway. The modification fixes the problem by checking the qi pointer before accessing it.


Signed-off-by: Feeling <u202012152@hust.edu.cn>