net/hinic: reassign the variable ret after sscanf() failure
driver inclusion
category: other
bugzilla: https://gitee.com/openeuler/kernel/issues/I6ZJSQ

Reference: N/A

---------------------

On line 767 of hinic_dbgtool_knl.c, when the function sscanf() fails,
the variable ret will be assigned an integer less than 1. Use 'ret < 1'
other than 'ret <= 0' in the 'if' judgment to change to fixed format.

And if the judgment is valid, the variable ret should be reassigned an
error code, which is -EINVAL in this case, for dbgtool_knl_init()
suggests an error code to help find bugs.

Signed-off-by: Ziliang Liao <saraday@hust.edu.cn>