feature: add support for GCC LTO compiler inclusion category: feature bugzilla: https://gitee.com/openeuler/kernel/issues/I73275
---------------------------------

This change adds support for GCC's Link Time Optimization (LTO).
With LTO, GCC will dump its internal representation (GIMPLE) to
disk, so that all the different compilation units that make up a
single executable can be optimized as a single module. This expands
the scope of inter-procedural optimizations to encompass the whole
program (or, rather, everything that is visible at link time). See:
https://gcc.gnu.org/onlinedocs/gccint/LTO.html

The Kconfig option CONFIG_LTO_GCC is implemented as a choice, which
defaults to LTO being disabled.

Our implementation is based on the work of Jiri Slaby, Andi Kleen and
Martin Liska.

Signed-off-by: Yancheng Li <liyancheng@huawei.com>