crypto/trng: Remove the automatic loading of the hisi_trng driver
On the current openEuler OS, when the OS is automatically started,
the hisi_trng_v2 device driver will be loaded because the algorithm
self-test needs to use random numbers.

However, the trng framework of community crypto requires that the
first loaded and used trng resource handle cannot be released, which
will cause hisi_trng_v2 to fail to unload the driver in subsequent
hisi trng module function processing.

The patch is modified by reducing the algorithm priority of the
hardware device driver. Not enabled in autotest and load. Only
reload this device driver for stand-alone use.

Signed-off-by: liulongfang <liulongfang@huawei.com>